patch three (v3/v4 upgrade): update docs to new format
