+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -8,23 +8,29 @@ assignees: ''
 ---
 
 **Describe the bug**
+
 A clear and concise description of what the bug is.
 
 **To Reproduce**
+
 Steps to reproduce the behavior:
 1. Run '...'
 2. See output '...'
 
 **Expected behavior**
+
 A clear and concise description of what you expected to happen.
 
 **Screenshots**
+
 If applicable, add screenshots to help explain your problem.
 
 **Versions (please complete the following information):**
+
  - OS: [e.g. iOS 12]
  - Rustc version [rustc --version]
  - Version [e.g. 22]
 
 **Additional context**
+
 Add any other context about the problem here.
