+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,3 +1,5 @@
+<!--
+
 Thanks for submitting a PR! We want to make contributing to Paseto as easy as possible.
 Please read these instructions carefully:
 
@@ -6,17 +8,21 @@ Please read these instructions carefully:
 - [ ] Match the **code formatting** of the rest of the codebase.
 - [ ] Make sure to **add tests** to help keep code coverage up.
 
-## Motivation (required) ##
+-->
+
+## Motivation <!-- (required) --> ##
 
 What existing problem does the pull request solve?
 
-## Test Plan (required) ##
+## Test Plan <!-- (required) --> ##
 
 A good test plan has the exact commands you ran and their output.
 
 If you have added code that should be tested, add tests.
 
+<!--
 ## Next Steps ##
 
 - Small pull requests are much easier to review and more likely to get merged. Make sure the PR does only one thing, otherwise please split it.
 - Make sure all **tests pass**, we will run automated tests, but you can run it yourself by running `cargo test`.
+-->
