+++ src/v2/public.rs
@@ -1,5 +1,4 @@
-//! An implementation of Paseto v2 "public" tokens, or tokens that
-//! are signed with a public/private key pair.
+//! ["Direct" use of public (signed but readable) tokens for V2 of Paseto.](https://github.com/paseto-standard/paseto-spec/blob/8b3fed8240e203b058649d01a82a8c412087bc87/docs/01-Protocol-Versions/Version2.md#sign)
 
 use crate::{
 	errors::{GenericError, PasetoError},
@@ -12,13 +11,13 @@ use ring::signature::{Ed25519KeyPair, UnparsedPublicKey, ED25519};
 
 const HEADER: &str = "v2.public.";
 
-/// Sign a "v2.public" paseto token.
+/// Sign a paseto token using `v2` of Paseto.
 ///
-/// Returns a result of a string if signing was successful.
+/// Returns a result of the token as a string if encryption was successful.
 ///
 /// # Errors
 ///
-/// If there was a failure signing the data.
+/// - If the calls to libsodium to sign your data fails.
 pub fn public_paseto(
 	msg: &str,
 	footer: Option<&str>,
@@ -46,13 +45,16 @@ pub fn public_paseto(
 	Ok(token)
 }
 
-/// Verifies a "v2.public" paseto token based on a given key pair.
+/// Verifies the signature of a paseto token using `v2` of Paseto, validating the footer.
 ///
-/// Returns the message if verification was successful.
+/// Returns the contents of the token as a string.
 ///
 /// # Errors
 ///
-/// If the string passed in was not a valid token, and did not have a correct footer.
+/// - If the token is not in the proper format: `v2.public.${signed_encoded_data}(.{optional_footer})?`
+/// - If the footer on the token did not match the footer passed in.
+/// - If we failed to validate the signature of the data.
+/// - If the data contained in the token was not valid utf-8.
 pub fn verify_paseto(
 	token: &str,
 	footer: Option<&str>,
